Insert type any for interval toggler
